Revert binary column
